% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{AllDataDscDensity}
\alias{AllDataDscDensity}
\title{Sort data by descending density}
\usage{
AllDataDscDensity(all_countries)
}
\arguments{
\item{all_countries}{A tibble of country data e.g. from
\code{geo.popn.density::GetAllData}}
}
\value{
A tibble with the same columns of the country data sorted in
descending order by population density by country and with the additional
column
\describe{
\item{Cum_Perc_Popn}{Cumulative percent of the population}
}
}
\description{
Sort data by descending density
}
