% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{ArrangeMSAsbyDensity}
\alias{ArrangeMSAsbyDensity}
\title{Sort metropolitan area data by descending population desnity}
\usage{
ArrangeMSAsbyDensity(msas)
}
\arguments{
\item{msas}{A tibble of msa data e.g. from
\code{geo.popn.density::GroupbyMSA}}
}
\value{
A tibble with the same columns as the msa data except as follows:
\code{Cum_Perc_Popn} cumulative sum of population percentage
}
\description{
Sort metropolitan area data by descending population desnity
}
