% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{ArrangeMSAsbyPopn}
\alias{ArrangeMSAsbyPopn}
\title{Sort metropolitan area data by population descending}
\usage{
ArrangeMSAsbyPopn(msas)
}
\arguments{
\item{msas}{A tibble of msa data e.g. from
\code{geo.popn.density::GroupbyMSA}}
}
\value{
A tibble with the same columns as the msa data except as follows:
\code{Rural} entries in the \code{MSA_Name} column placed last
\code{Cum_Perc_Popn} cumulative sum of population percentage
}
\description{
Sort metropolitan area data by population descending
}
