% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{au}
\alias{au}
\title{Data for Australian
Population and population densities at the ABS SA2 level}
\format{
A data frame with 2288 rows and 14 variables:
\describe{
\item{Country_Code}{Upper case ISO 3166 code}
\item{State_Code}{ABS state code}
\item{State_Abbr}{2 or 3 letter state abbreviation}
\item{State_Name}{State name}
\item{County_Code}{County code - \code{NA} for AU data}
\item{County_Name}{County name - \code{NA} for AU data}
\item{MSA_Code}{ABS significant urban area code}
\item{MSA_Name}{ABS significant urban area name}
\item{Metro_Micro}{\code{Metro} if a metropolitan area, \code{NA} otherwise}
\item{SA_Code}{ABS statistical area level 2 code}
\item{SA_Name}{ABS statistical area level 2 name}
\item{Popn_Est}{Population estimate, currenlty for 2021}
\item{Area_km2}{Area of the relevant statistical area}
\item{Density}{Derived density}
}
}
\usage{
au
}
\description{
This dataset was produced from Australian Bureau of Statistics data
available from  \url{https://www.abs.gov.au/}. See the references section
in the readme for further detail on Australian data and definitions.
}
\keyword{datasets}
