% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{CalcAvgWDensities}
\alias{CalcAvgWDensities}
\title{Calculates average weighted densities for each country}
\usage{
CalcAvgWDensities(all_countries)
}
\arguments{
\item{all_countries}{A tibble of country data e.g. from
\code{geo.popn.density::GetAllData}}
}
\value{
A tibble columns as follows:
\describe{
\item{Country_Code}{Upper case ISO 3166 code}
\item{avgw_density}{Average weighted population density in km^2}
\item{method}{Calculation density}
}
}
\description{
Calculates using the arithmetic, geometric, and median methods
}
