% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcCumPercent.R
\name{CalcCumPercent}
\alias{CalcCumPercent}
\title{Sort data by descending metric and add cumulative percentage column}
\usage{
CalcCumPercent(
  data,
  cum_pct = "Cum_Perc",
  metric = NULL,
  sort_by = NULL,
  group_by = NULL
)
}
\arguments{
\item{data}{A data frame or equivalent (e.g. tibble) to sort}

\item{cum_pct}{The name for the new; default is "Cum_Perc"}

\item{metric}{The column in \code{data} that contains the metric (e.g. population)
to calculate the cumulative percentage of. Required, will exit if \code{NULL}.}

\item{sort_by}{The column in \code{data} to sort in descending order e.g. density.
Required, will exit if if \code{NULL}.}

\item{group_by}{A character vector containing the column names in \code{data}
to group and calculate the cumulative percentages for e.g.
\verb{c('Country_Code', 'State_Name)}. If specified will remove existing grouping.
If \code{NULL} then no grouping will occur.}
}
\value{
A data frame of the same type as passed with data sorted in
descending order by \code{sort_by} and grouped if specified with the additional
column
\describe{
\item{cum_pct}{Cumulative percent of the given metric}
}
}
\description{
If specified, will calculate cumulative percentage for each group e.g. by country
}
\details{
Mainly used to calculate the cumulative percentage population for data
sorted by population density.
}
