% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcRural.R
\name{ExcRural}
\alias{ExcRural}
\title{Sort data by descending metric and add cumulative percentage column}
\usage{
ExcRural(data, msa_name, msa_code)
}
\arguments{
\item{data}{A data frame or equivalent (e.g. tibble) to sort}

\item{msa_name}{Column name for the metroplitan area names}

\item{msa_code}{Column name for the metroplitan area codes}
}
\value{
A data frame of the same type as passed
}
\description{
If specified, will calculate cumulative percentage for each group e.g. by country
}
\details{
Mainly used to calculate the cumulative percentage population for data
sorted by population density.
}
