% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMedian.R
\name{FindMedian}
\alias{FindMedian}
\title{Find the median row by specified metric}
\usage{
FindMedian(data, cum_pct = NULL)
}
\arguments{
\item{data}{A data frame or equivalent (e.g. tibble) to find the median row}

\item{cum_pct}{The column in \code{data} that contains the cumulative
percentage to sort by. If \code{NULL} then will create and add the column to
to \code{data} with the default name of \code{Cum_Perc}.}
}
\value{
A data frame of the same type as passed with the median row for each
group if \code{data} is grouped
}
\description{
Mainly used to find the median population density for each country. Note,
\code{base::median()} finds the median value by sorting the data frame and
returning the median (i.e. middle) row. This function is intended to be used
to find the row where the cumulative percentage (typically of population)
is >= 50\%.
}
