% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{FindMedianMSAbyPopulation}
\alias{FindMedianMSAbyPopulation}
\title{Find the median metropolitan area by population density}
\usage{
FindMedianMSAbyPopulation(msas_by_popn)
}
\arguments{
\item{msas_by_popn}{A tibble of msa data e.g. from
\code{geo.popn.density::ArrangeMSAsbyPopn}}
}
\value{
A tibble with columns as follows:
\describe{
\item{Country_Code}{Cumulative percent of the population}
\item{MSA_Name}{Metropolitan area name}
\item{Popn_Est}{Population estimate}
}
}
\description{
Find the median metropolitan area by population density
}
