% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{FindMedianMSAOccurencebyDensity}
\alias{FindMedianMSAOccurencebyDensity}
\title{Find the median metropolitan area occurence by population density}
\usage{
FindMedianMSAOccurencebyDensity(msas_by_density)
}
\arguments{
\item{msas_by_density}{A tibble of msa data e.g. from
\code{geo.popn.density::ArrangeMSAsbyDensity}}
}
\value{
A tibble with columns as follows:
\describe{
\item{Country_Code}{Cumulative percent of the population}
\item{MSA_Name}{Metropolitan area name}
\item{Popn_Est}{Population estimate}
}
}
\description{
Find the median metropolitan area occurence by population density
}
