% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{FindMedianSAbyDensity}
\alias{FindMedianSAbyDensity}
\title{Find the median statistical area by population density}
\usage{
FindMedianSAbyDensity(allc_by_density)
}
\arguments{
\item{allc_by_density}{A tibble of country data e.g. from
\code{geo.popn.density::AllDataDscDensity}}
}
\value{
A tibble with columns as follows:
\describe{
\item{Country_Code}{Cumulative percent of the population}
\item{MSA_Name}{Metropolitan area name}
\item{SA_Code}{Statistical area code}
\item{Popn_Est}{Population estimate}
}
}
\description{
Find the median statistical area by population density
}
