% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{GroupbyMSA}
\alias{GroupbyMSA}
\title{Group the statistical area level by metropolitan area}
\usage{
GroupbyMSA(all_countries)
}
\arguments{
\item{all_countries}{A tibble of country data e.g. from
\code{geo.popn.density::GetAllData}}
}
\value{
A tibble with the same columns as the all country data
}
\description{
Group the statistical area level by metropolitan area
}
