% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupData.R
\name{GroupData}
\alias{GroupData}
\title{Group the statistical area data by desired grouping}
\usage{
GroupData(
  data,
  ppn_est = "ppn_est",
  area = "area",
  ppn_pct = "ppn_pct",
  density = "density",
  group_by = NULL
)
}
\arguments{
\item{data}{A tibble of country data e.g. from \code{geo.popn.density::GetAllData}}

\item{ppn_est}{Column name that contains the population estimate; default is "ppn_est"}

\item{area}{Column name that contains the area; default is "area"}

\item{ppn_pct}{Column name that contains the population percentage for each area; default is "ppn_pct"}

\item{density}{Column name that contains the density; default is "density"}

\item{group_by}{A character vector containing the column names in \code{data}
to group and calculate the cumulative percentages for e.g.
\verb{c('Country_Code', 'State_Name)}. If \code{NULL} then no grouping will occur.}
}
\value{
A tibble with the same columns as the all country data
}
\description{
Group the statistical area data by desired grouping
}
