% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAllDataByDensity.R
\name{PlotAllDataByDensity}
\alias{PlotAllDataByDensity}
\title{Plot pareto chart of density vs. cumulative population}
\usage{
PlotAllDataByDensity(
  allc_by_density,
  density = "density",
  ccode = "country_code",
  cum_pct = "cum_pct",
  mdn_sa
)
}
\arguments{
\item{allc_by_density}{A tibble of country data e.g. from \code{geo.popn.density::AllDataDscDensity}}

\item{density}{Column name that contains the density; default is "density"}

\item{ccode}{Column name that contains the country code; default is "country_code"}

\item{cum_pct}{Column name that contains the cumulative percentage; default is "cum_pct"}

\item{mdn_sa}{Tibble containing median entries e.g. from \code{geo.popn.density::FindMedian}}
}
\description{
Plot pareto chart of density vs. cumulative population
}
