% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotAvgWDensities}
\alias{PlotAvgWDensities}
\title{Plot average weighted densities for each country by method}
\usage{
PlotAvgWDensities(avgw_densities)
}
\arguments{
\item{avgw_densities}{A tibble of country data e.g. from
\code{geo.popn.density::CalcAvgWDensities}
\describe{
\item{Country_Code}{Upper case ISO 3166 code}
\item{avgw_density}{Average weighted population density in km^2}
\item{method}{Calculation density}
}}
}
\description{
Plot average weighted densities for each country by method
}
