% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDensities.R
\name{PlotDensities}
\alias{PlotDensities}
\title{Plot densities for each country by method}
\usage{
PlotDensities(densities, density = "density", ccode = "country_code")
}
\arguments{
\item{densities}{A tibble of country data e.g. from \code{geo.popn.density::CalcDensities}}

\item{density}{Column name that contains the density; default is "density"}

\item{ccode}{Column name that contains the country code; default is "country_code"
\describe{
\item{Country_Code}{Upper case ISO 3166 code}
\item{Density}{Population density in km^2}
\item{Method}{Calculation method}
}}
}
\description{
Plot densities for each country by method
}
