% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotMSAsByDensity}
\alias{PlotMSAsByDensity}
\title{Plot average weighted densities for each country by method}
\usage{
PlotMSAsByDensity(
  msas_by_density,
  density = "density",
  ccode = "country_code",
  cum_pct = "cum_pct"
)
}
\arguments{
\item{msas_by_density}{A tibble of country data e.g. from \code{geo.popn.density::CalcAvgWDensities}}

\item{density}{Column name that contains the density; default is "density"}

\item{ccode}{Column name that contains the country code; default is "country_code"}

\item{cum_pct}{Column name that contains the cumulative percentage; default is "cum_pct"
\describe{
\item{Country_Code}{Upper case ISO 3166 code}
\item{avgw_density}{Average weighted population density in km^2}
\item{method}{Calculation density}
}}
}
\description{
Plot average weighted densities for each country by method
}
