% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMSAsByPopn.R
\name{PlotMSAsByPopn}
\alias{PlotMSAsByPopn}
\title{Plot average weighted densities for each country by method}
\usage{
PlotMSAsByPopn(
  msas_by_popn,
  ppn_est = "ppn_est",
  ccode = "country_code",
  cum_pct = "cum_pct"
)
}
\arguments{
\item{msas_by_popn}{A tibble of country data e.g. from \code{geo.popn.density::CalcAvgWDensities}}

\item{ppn_est}{Column name that contains the population estimate; default is "ppn_est"}

\item{ccode}{Column name that contains the country code; default is "country_code"}

\item{cum_pct}{Column name that contains the cumulative percentage; default is "cum_pct"
\describe{
\item{Country_Code}{Upper case ISO 3166 code}
\item{Density}{Average weighted population density in km^2}
\item{Method}{Calculation density}
}}
}
\description{
Plot average weighted densities for each country by method
}
