% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ShowCountryTotals}
\alias{ShowCountryTotals}
\title{Show table with median result for each group}
\usage{
ShowCountryTotals(data)
}
\arguments{
\item{data}{A tibble of median data e.g. from \code{geo.popn.density::FindMedian}
\describe{
\item{Country_Code}{Upper case ISO 3166 code}
\item{Popn_Est}{Population estimate}
\item{Area_km2}{Area in km^2}
}}
}
\value{
A reactable formatted table
}
\description{
Show table with median result for each group
}
