% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowTable.R
\name{ShowTable}
\alias{ShowTable}
\title{Show table in html div with css class}
\usage{
ShowTable(tbl, css_class)
}
\arguments{
\item{tbl}{Table to show}

\item{css_class}{String name of css class}
}
\description{
Show table in html div with css class
}
