% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortMSAsbyDensity.R
\name{SortMSAsbyDensity}
\alias{SortMSAsbyDensity}
\title{Sort metropolitan area data by descending population density}
\usage{
SortMSAsbyDensity(
  msas,
  cum_pct = "cum_pct",
  density = "density",
  ppn_pct = "ppn_pct",
  group_by = NULL
)
}
\arguments{
\item{msas}{A tibble of msa data e.g. from \code{geo.popn.density::GroupbyMSA}}

\item{cum_pct}{Column name that contains the cumulative percentage; default is "cum_pct"}

\item{density}{Column name that contains the density estimate; default is "density"}

\item{ppn_pct}{Column name that contains the population percentage for each area; default is "ppn_pct"}

\item{group_by}{A character vector containing the column names in \code{data}
to group and sort by. Specifying this will remove all other grouping.
If \code{NULL} then no grouping will occur.}
}
\value{
A tibble with the same columns as the msa data except as follows:
\describe{
\item{Cum_Perc_Popn}{Cumulative sum of population percentage}
}
}
\description{
Sort metropolitan area data by descending population density
}
