% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortMSAsbyPopn.R
\name{SortMSAsbyPopn}
\alias{SortMSAsbyPopn}
\title{Sort metropolitan area data by descending population}
\usage{
SortMSAsbyPopn(
  msas,
  cum_pct = "cum_pct",
  ppn_est = "ppn_est",
  ppn_pct = "ppn_pct",
  msa_name = "msa_name",
  group_by = NULL
)
}
\arguments{
\item{msas}{A tibble of MSAs data e.g. from \code{geo.popn.density::GroupbyMSA}}

\item{cum_pct}{The name for the new cumulative percentage column; default is \code{cum_pct}}

\item{ppn_est}{Column name that contains the population estimate; default is "ppn_est"}

\item{ppn_pct}{Column name that contains the population percentage for each area; default is "ppn_pct"}

\item{msa_name}{Column name that contains the MSA name; default is "msa_name"}

\item{group_by}{A character vector containing the column names in \code{data}
to group and sort by. Specifying this will remove all other grouping.
If \code{NULL} then no grouping will occur.}
}
\value{
A tibble with the same columns as the msa data except as follows:
\describe{
\item{Rural}{Entries in the \code{MSA_Name} column placed last in the tibble}
\item{Cum_Perc_Popn}{Cumulative sum of population percentage}
}
}
\description{
Sort metropolitan area data by descending population
}
