% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMedian.R
\name{TableMedian}
\alias{TableMedian}
\title{Show table with median result for each group}
\usage{
TableMedian(data, show_cols)
}
\arguments{
\item{data}{A tibble of median data e.g. from \code{geo.popn.density::FindMedian}}

\item{show_cols}{List that contains the columns to display from
\describe{
\item{Country_Code}{Upper case ISO 3166 code}
\item{MSA_Name}{Metropolitan satistical area name}
\item{SA_Code}{Statistical area code}
\item{Popn_Est}{Population estimate}
\item{Area_km2}{Area in km^2}
\item{Density}{Population density in km^2}
\item{Cum_Popn_Perc}{Cumulative population percent}
}}
}
\value{
A reactable formatted table
}
\description{
Show table with median result for each group
}
