% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableSummStats.R
\name{TableSummStats}
\alias{TableSummStats}
\title{Show table of summary statistics for each group by country}
\usage{
TableSummStats(data, group_by = NULL, stat_col, dflt_dp = 0)
}
\arguments{
\item{data}{A tibble of median data e.g. from \code{geo.popn.density::FindMedian}
\describe{
\item{Country_Code}{Upper case ISO 3166 code}
\item{Popn_Est}{Population estimate}
\item{Area_km2}{Area in km^2}
\item{Optional: Additional group columns}{e.g. MSA_Name, State_Name}
}}

\item{group_by}{A character vector containing the column names in \code{data}
to group and calculate the cumulative percentages for e.g.
\code{c("State_Name")}. If \code{NULL} then only grouping by \code{Country_Code} will occur.}

\item{stat_col}{Column name to calculate the stats on e.g. \code{area}}

\item{dflt_dp}{Default number of decimal points to show}
}
\value{
A reactable formatted table
}
\description{
Show table of summary statistics for each group by country
}
