% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{GetAllData}
\alias{GetAllData}
\title{Combines and returns all country data}
\usage{
GetAllData()
}
\value{
A tibble with the same columns of the country data except as follows:
\describe{
\item{MSA_Name}{\code{NA} entries in this column are replaced with \code{Rural}}
\item{Perc_Popn}{New column with percentage of each countries' population}
}
}
\description{
Combines and returns all country data
}
