% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotAllDataByDensity}
\alias{PlotAllDataByDensity}
\title{Plot pareto chart of density vs. cumulative population}
\usage{
PlotAllDataByDensity(allc_by_density)
}
\arguments{
\item{allc_by_density}{A tibble of country data e.g. from
\code{geo.popn.density::AllDataDscDensity}}
}
\description{
Plot pareto chart of density vs. cumulative population
}
